# Service optimization: Database indexes, query optimization, and vectorized processing